fix ndppp bug

+++ autocal/pipeline-cal.py
@@ -75,6 +75,7 @@ else:
     logging.info('Copy data - no averaging...')
     for ms in mss:
         msout = ms.replace('.MS','-avg.MS').split('/')[-1]
+        if os.path.exists(msout): continue
         os.system('cp -r '+ms+' '+msout)
 
 mss = sorted(glob.glob('*.MS'))
@@ -130,7 +131,7 @@ logging.info('Calibrating...')
 for ms in mss:
     check_rm(ms+'/instrument')
     s.add('NDPPP '+parset_dir+'/NDPPP-sol.parset msin='+ms+' sol.sourcedb='+sourcedb+' sol.sources='+patch, log=ms+'_sol1.log', cmd_type='NDPPP')
-s.run(check=True, max_threads=s.max_threads/6)
+s.run(check=True)
 
 # Prepare and run losoto
 check_rm('globaldb')
@@ -173,7 +174,7 @@ for i, ms in enumerate(mss):
 logging.info('Cross delay correction...')
 for ms in mss:
     s.add('NDPPP '+parset_dir+'/NDPPP-cor.parset msin='+ms+' cor.parmdb='+ms+'/instrument-cd cor.correction=gain', log=ms+'_corCD.log', cmd_type='NDPPP')
-s.run(check=True, max_threads=s.max_threads/6)
+s.run(check=True)
 
 # Convert to circular CORRECTED_DATA -> CORRECTED_DATA
 logging.info('Converting to circular...')
@@ -192,7 +193,7 @@ logging.info('Calibrating...')
 for ms in mss:
     check_rm(ms+'/instrument')
     s.add('NDPPP '+parset_dir+'/NDPPP-sol.parset msin='+ms+' sol.sourcedb='+sourcedb+' sol.sources='+patch, log=ms+'_sol2.log', cmd_type='NDPPP')
-s.run(check=True, max_threads=s.max_threads/6)
+s.run(check=True)
 
 # Prepare and run losoto
 check_rm('globaldb')
@@ -250,7 +251,7 @@ s.run(check=True)
 logging.info('Faraday rotation correction...')
 for ms in mss:
     s.add('NDPPP '+parset_dir+'/NDPPP-cor.parset msin='+ms+' cor.parmdb='+ms+'/instrument-fr cor.correction=RotationMeasure', log=ms+'_corFR.log', cmd_type='NDPPP')
-s.run(check=True, max_threads=s.max_threads/6)
+s.run(check=True)
 
 # Convert to circular CORRECTED_DATA -> CORRECTED_DATA
 # NOTE: in linear
@@ -270,7 +271,7 @@ logging.info('Calibrating...')
 for ms in mss:
     check_rm(ms+'/instrument')
     s.add('NDPPP '+parset_dir+'/NDPPP-sol.parset msin='+ms+' sol.sourcedb='+sourcedb+' sol.sources='+patch, log=ms+'_sol3.log', cmd_type='NDPPP')
-s.run(check=True, max_threads=s.max_threads/6)
+s.run(check=True)
 
 # Prepare and run losoto
 check_rm('globaldb') # remove it as it was used for the fr
